+++ compiler/src/org.graalvm.compiler.truffle.runtime/src/org/graalvm/compiler/truffle/runtime/FrameWithoutBoxing.java
@@ -119,7 +119,8 @@ public final class FrameWithoutBoxing implements VirtualFrame, MaterializedFrame
     @Override
     public MaterializedFrame materialize() {
         if (!getFrameDescriptor().canMaterialize()) {
-            throw new IllegalArgumentException("Cannot materialize a frame for which materialization was explicitly disallowed.");
+            CompilerDirectives.transferToInterpreterAndInvalidate();
+            throw new UnsupportedOperationException("Cannot materialize a frame for which materialization was explicitly disallowed.");
         }
         ((GraalTruffleRuntime) Truffle.getRuntime()).markFrameMaterializeCalled(descriptor);
         return this;