+++ compiler/src/org.graalvm.compiler.truffle.test/src/org/graalvm/compiler/truffle/test/BytecodeOSRNodeTest.java
@@ -79,7 +79,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
      */
     @Test
     public void testSimpleInterpreterLoop() {
-        RootNode rootNode = new Program(new InfiniteInterpreterLoop(), new FrameDescriptor(false));
+        RootNode rootNode = new Program(new InfiniteInterpreterLoop(), new FrameDescriptor(null, false));
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
         // Interpreter invocation should be OSR compiled and break out of the interpreter loop.
         Assert.assertEquals(42, target.call());
@@ -90,7 +90,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
      */
     @Test
     public void testFixedIterationLoop() {
-        FrameDescriptor frameDescriptor = new FrameDescriptor(false);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, false);
         FixedIterationLoop osrNode = new FixedIterationLoop(frameDescriptor);
         RootNode rootNode = new Program(osrNode, frameDescriptor);
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
@@ -102,7 +102,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
      */
     @Test
     public void testFixedIterationLoopBelowThreshold() {
-        FrameDescriptor frameDescriptor = new FrameDescriptor(false);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, false);
         FixedIterationLoop osrNode = new FixedIterationLoop(frameDescriptor);
         RootNode rootNode = new Program(osrNode, frameDescriptor);
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
@@ -115,21 +115,21 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
     @Test
     public void testMultipleLoops() {
         // Each loop runs for osrThreshold + 1 iterations, so the first loop should trigger OSR.
-        FrameDescriptor frameDescriptor = new FrameDescriptor(false);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, false);
         TwoFixedIterationLoops osrNode = new TwoFixedIterationLoops(frameDescriptor);
         RootNode rootNode = new Program(osrNode, frameDescriptor);
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
         Assert.assertEquals(TwoFixedIterationLoops.OSR_IN_FIRST_LOOP, target.call(osrThreshold + 1));
 
         // Each loop runs for osrThreshold/2 + 1 iterations, so the second loop should trigger OSR.
-        frameDescriptor = new FrameDescriptor(false);
+        frameDescriptor = new FrameDescriptor(null, false);
         osrNode = new TwoFixedIterationLoops(frameDescriptor);
         rootNode = new Program(osrNode, frameDescriptor);
         target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
         Assert.assertEquals(TwoFixedIterationLoops.OSR_IN_SECOND_LOOP, target.call(osrThreshold / 2 + 1));
 
         // Each loop runs for osrThreshold/2 iterations, so OSR should not get triggered.
-        frameDescriptor = new FrameDescriptor(false);
+        frameDescriptor = new FrameDescriptor(null, false);
         osrNode = new TwoFixedIterationLoops(frameDescriptor);
         rootNode = new Program(osrNode, frameDescriptor);
         target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
@@ -141,7 +141,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
      */
     @Test
     public void testFailedCompilation() {
-        FrameDescriptor frameDescriptor = new FrameDescriptor(false);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, false);
         UncompilableFixedIterationLoop osrNode = new UncompilableFixedIterationLoop(frameDescriptor);
         RootNode rootNode = new Program(osrNode, frameDescriptor);
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
@@ -155,7 +155,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
      */
     @Test
     public void testInvalidateOnNodeReplaced() {
-        FrameDescriptor frameDescriptor = new FrameDescriptor(false);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, false);
         Node childToReplace = new Node() {
         };
         FixedIterationLoop osrNode = new FixedIterationLoopWithChild(frameDescriptor, childToReplace);
@@ -185,7 +185,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
      */
     @Test
     public void testOSRWithMaterializeableFrame() {
-        FrameDescriptor frameDescriptor = new FrameDescriptor(true);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, true);
         FixedIterationLoop osrNode = new FixedIterationLoop(frameDescriptor);
         RootNode rootNode = new Program(osrNode, frameDescriptor);
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
@@ -207,7 +207,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
                         "engine.BackgroundCompilation", Boolean.TRUE.toString() // override defaults
         );
         InfiniteInterpreterLoop osrNode = new InfiniteInterpreterLoop();
-        RootNode rootNode = new Program(osrNode, new FrameDescriptor(false));
+        RootNode rootNode = new Program(osrNode, new FrameDescriptor(null, false));
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
         Assert.assertEquals(42, target.call());
         BytecodeOSRMetadata osrMetadata = osrNode.getGraalOSRMetadata();
@@ -221,7 +221,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
      */
     @Test
     public void testStackTraceHidesOSRCallTarget() {
-        FrameDescriptor frameDescriptor = new FrameDescriptor(false);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, false);
         CheckStackWalkCallTarget osrNode = new CheckStackWalkCallTarget(frameDescriptor);
         RootNode rootNode = new Program(osrNode, frameDescriptor);
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
@@ -233,7 +233,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
      */
     @Test
     public void testStackTraceUsesOSRFrame() {
-        FrameDescriptor frameDescriptor = new FrameDescriptor(false);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, false);
         CheckStackWalkFrame osrNode = new CheckStackWalkFrame(frameDescriptor);
         RootNode rootNode = new Program(osrNode, frameDescriptor);
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
@@ -247,7 +247,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
      */
     @Test
     public void testStackTraceUsesNewestOSRFrame() {
-        FrameDescriptor frameDescriptor = new FrameDescriptor(false);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, false);
         CheckStackWalkFrameNested osrNode = new CheckStackWalkFrameNested(frameDescriptor);
         RootNode rootNode = new Program(osrNode, frameDescriptor);
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
@@ -264,7 +264,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
      */
     @Test
     public void testGetCallerFrameSkipsOSR() {
-        FrameDescriptor frameDescriptor = new FrameDescriptor(false);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, false);
         CheckGetCallerFrameSkipsOSR osrNode = new CheckGetCallerFrameSkipsOSR(frameDescriptor);
         RootNode rootNode = new Program(osrNode, frameDescriptor);
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
@@ -279,7 +279,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
      */
     @Test
     public void testFrameTransfer() {
-        FrameDescriptor frameDescriptor = new FrameDescriptor(false);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, false);
         RootNode rootNode = new Program(new FrameTransferringNode(frameDescriptor), frameDescriptor);
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
         Assert.assertEquals(42, target.call());
@@ -292,7 +292,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
      */
     @Test
     public void testFrameTransferWithTagUpdate() {
-        FrameDescriptor frameDescriptor = new FrameDescriptor(false);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, false);
         RootNode rootNode = new Program(new FrameTransferringNodeWithTagUpdate(frameDescriptor), frameDescriptor);
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
         Assert.assertEquals(42, target.call());
@@ -331,7 +331,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
     @Test
     public void testOSRInBytecodeLoop() {
         // osrThreshold + 1 back-edges -> compiled
-        FrameDescriptor frameDescriptor = new FrameDescriptor(false);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, false);
         BytecodeNode bytecodeNode = new BytecodeNode(3, frameDescriptor, tripleInput1);
         RootNode rootNode = new Program(bytecodeNode, frameDescriptor);
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
@@ -346,7 +346,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
     @Test
     public void testNoOSRInBytecodeLoop() {
         // osrThreshold back-edges -> not compiled
-        FrameDescriptor frameDescriptor = new FrameDescriptor(false);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, false);
         BytecodeNode bytecodeNode = new BytecodeNode(3, frameDescriptor, tripleInput1);
         RootNode rootNode = new Program(bytecodeNode, frameDescriptor);
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
@@ -362,7 +362,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
         // computes osrThreshold * 2
         // Inner loop contributes 1 back-edge, so each outer loop contributes 2 back-edges, and
         // the even-valued osrThreshold gets hit by the outer loop back-edge.
-        FrameDescriptor frameDescriptor = new FrameDescriptor(false);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, false);
         BytecodeNode bytecodeNode = new BytecodeNode(4, frameDescriptor, multiplyInputs);
         RootNode rootNode = new Program(bytecodeNode, frameDescriptor);
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);
@@ -380,7 +380,7 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
         // Inner loop contributes osrThreshold-2 back-edges, so the first outer loop contributes
         // osrThreshold-1 back-edges, then the next back-edge (which triggers OSR) is from the inner
         // loop.
-        FrameDescriptor frameDescriptor = new FrameDescriptor(false);
+        FrameDescriptor frameDescriptor = new FrameDescriptor(null, false);
         BytecodeNode bytecodeNode = new BytecodeNode(4, frameDescriptor, multiplyInputs);
         RootNode rootNode = new Program(bytecodeNode, frameDescriptor);
         OptimizedCallTarget target = (OptimizedCallTarget) runtime.createCallTarget(rootNode);