+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/impl/DefaultVirtualFrame.java
@@ -42,6 +42,7 @@ package com.oracle.truffle.api.impl;
 
 import java.util.Arrays;
 
+import com.oracle.truffle.api.CompilerDirectives;
 import com.oracle.truffle.api.TruffleRuntime;
 import com.oracle.truffle.api.frame.FrameDescriptor;
 import com.oracle.truffle.api.frame.FrameSlot;
@@ -83,7 +84,8 @@ final class DefaultVirtualFrame implements VirtualFrame {
     @Override
     public MaterializedFrame materialize() {
         if (!getFrameDescriptor().canMaterialize()) {
-            throw new IllegalArgumentException("Cannot materialize a frame for which materialization was explicitly disallowed.");
+            CompilerDirectives.transferToInterpreterAndInvalidate();
+            throw new UnsupportedOperationException("Cannot materialize a frame for which materialization was explicitly disallowed.");
         }
         return new DefaultMaterializedFrame(this);
     }